[FIX]:Video entity 중복되어 변경
