+++ src/main/java/com/semtleWebGroup/youtubeclone/domain/video/repository/VideoRepository.java
@@ -1,6 +1,8 @@
 package com.semtleWebGroup.youtubeclone.domain.video.repository;
+
 import com.semtleWebGroup.youtubeclone.domain.video.domain.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.List;
 
 public interface VideoRepository extends JpaRepository<Video, Integer> {
